# Conflicts:
#	server/src/categories/models.py

+++ server/src/currencies/services/currency_toolkit.py
@@ -1,11 +1,12 @@
-from accounts.models import User, account
+  
+from conftest import account
+from accounts.models import User, Account
 from currencies.models import Currency
 
 
 class CurrencyToolkit:
-    # Add sort by user
     @classmethod
-    def get_currency(cls, user, name) -> Currency:  
+    def get_currency(cls, user: User, name: str) -> Currency:  
         qs = Currency.objects.filter(name=name, public=True)
         if qs.exists():
             return qs.first()
@@ -17,7 +18,7 @@ class CurrencyToolkit:
         raise Currency.DoesNotExist()
 
     @classmethod
-    def create_currency(cls, name, code, account):
+    def create_currency(cls, name: str, code: str, account: Account):
         from currencies.services import CurrencyCreator
         currency = CurrencyCreator(
             name=name,
@@ -28,20 +29,23 @@ class CurrencyToolkit:
         return currency
 
     @classmethod
-    def delete_currency(cls, currency_id):
-        currency = cls.get_currency(currency_id=currency_id)
-        currency.delete()
-        return True
+    def delete_currency(cls, currency_id: int):
+        if Currency.objects.filter(id=id, public=False, account=account).exists():
+            currency = Currency.objects.get(id=currency_id)
+            currency.delete()
+            return True
+        else:
+            raise Currency.DoesNotExist()
+
 
     @classmethod
-    def edit_currency(cls, id, new_name, new_code):
+    def edit_currency(cls, id: int, new_name: str, new_code: str, account: Account):
         from currencies.services import CurrencyEditor
-        return CurrencyEditor(
-            id=id,
-            new_name=new_name,
-            new_code=new_code
-        )()
-
-    @classmethod
-    def get_accounts_currencies(cls, account):
-        pass
+        if Currency.objects.filter(id=id, public=False, account=account).exists():
+            return CurrencyEditor(
+                id=id,
+                new_name=new_name,
+                new_code=new_code
+            )()
+        else:
+            raise Currency.DoesNotExist()
